Switched evaluation value type to int32

+++ src/engine.rs
@@ -6,5 +6,5 @@ pub trait Engine {
     fn stop(&mut self);
     fn update(&mut self, fen: Position, moves: Vec<ChessMove>);
     fn restart(&mut self);
-    fn evaluate(&self)->f32;
+    fn evaluate(&self)->i32;
 }
